requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
comm==0.1.4
DateTime==5.2
debugpy==1.8.0
decorator==5.1.1
discord.py==2.3.2
exceptiongroup==1.1.3
executing==2.0.1
frozenlist==1.4.0
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
jupyter_client==8.5.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pexpect==4.8.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
urllib3==2.0.7
wcwidth==0.2.9
yarl==1.9.2
zope.interface==6.1
